(Simulation.java) Made sure in all computeMortality cases that computePredationSuccess() is called just after predation. Added the mortality rates in the DEBUG file nDead.csv of computeMortality function.
git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@264 45b28200-7c74-4d1a-96bc-342942c5875a

+++ Osmose/src/fr/ird/osmose/Indicators.java
@@ -198,8 +198,8 @@ public class Indicators {
             pr.print(time);
             for (int i = 0; i < getSimulation().getNbSpecies(); i++) {
                 pr.print(";");
-                //pr.print(variable[i]);
-                pr.print((long) variable[i]);
+                pr.print(variable[i]);
+                //pr.print((long) variable[i]);
                 //System.out.println(filename + " " + time + " spec" + i + " " + variable[i]);
             }
             pr.println();
@@ -249,8 +249,8 @@ public class Indicators {
             for (int i = 0; i < getSimulation().getNbSpecies(); i++) {
                 for (int j = 0; j < variable[i].length; j++) {
                     pr.print(";");
-                    //pr.print(variable[i][j]);
-                    pr.print((long) variable[i][j]);
+                    pr.print(variable[i][j]);
+                    //pr.print((long) variable[i][j]);
                     //System.out.println(filename + " " + time + " spec" + i + " " + variable[i]);
                 }
             }
